pme: hexedit fingerprint HAL to send KEY_HOMEPAGE
It used to send KEY_WAKEUP (0x8F) instead of KEY_HOMEPAGE (0xAC).
CyanogenMod doesn't handle KEY_WAKEUP as a home button.
This was possible with the help of IDA Pro. After disassembling,
I looked for instances of 0x8F in the disassembly, then reassembled
replacing 0x8F with 0xAC. This is terrible hackery, I know.

Change-Id: I098631e7ffe1fb06ec5d7c32f446afa38b1b15c0